LayoutTests/imported/w3c/web-platform-tests/accname/aria-owns-expected.txt

Accname tests for assorted combinations of aria-owns, aria-hidden, and host language hiding techniques

Text enclosed in a dashed red outline is aria-hidden="true".

Bold text with a yellow background is the target of another element’s aria-owns.

A button using aria-owns to specify its label


Play Pause
A link using aria-owns to concatenate extra text

World Wide Web Consortium
(opens in a new window)
Ignore aria-owns when on an element that is hidden from all users

treasure

Computed name of parent heading persists when aria-owns fails to relocate its contents

sand

a pearl

Computed name of parent heading excludes content relocated by aria-owns

The dog that caught the

Speeding car

Ignore aria-owns when on an element with aria-hidden

This vampire can’t see its... reflection

Ignore aria-owns when it targets an element that is hidden from all users

World Wide Web Consortium
Ignore aria-owns when it targets an element with an ancestor that is hidden from all users

World Wide Web Consortium
aria-hidden remains in effect on an element after it is relocated using aria-owns

World Wide Web Consortium
(opens in a new window)

FAIL A button using aria-owns to specify its label assert_equals: <button class="ex-label" data-testname="A button using aria-owns to specify its label" data-expectedlabel="Play" aria-owns="play">
    <div aria-hidden="true">
      <span id="play">Play</span>
      <span id="pause">Pause</span>
    </div>
  </button> expected "Play" but got ""
FAIL A link using aria-owns to concatenate extra text assert_equals: <a class="ex-label" data-testname="A link using aria-owns to concatenate extra text" data-expectedlabel="World Wide Web Consortium (opens in a new window)" href="https://www.w3.org/" target="_blank" aria-owns="new-window-warning-1">
    World Wide Web Consortium
  </a> expected "World Wide Web Consortium (opens in a new window)" but got "World Wide Web Consortium"
PASS Ignore aria-owns when on an element that is hidden from all users
PASS Computed name of parent heading persists when aria-owns fails to relocate its contents
FAIL Computed name of parent heading excludes content relocated by aria-owns assert_equals: <h4 class="ex-label" data-testname="Computed name of parent heading excludes content relocated by aria-owns" data-expectedlabel="Speeding">
    Speeding
    <mark id="car">car</mark>
  </h4> expected "Speeding" but got "Speeding car"
PASS Ignore aria-owns when on an element with aria-hidden
PASS Ignore aria-owns when it targets an element that is hidden from all users
PASS Ignore aria-owns when it targets an element with an ancestor that is hidden from all users
PASS aria-hidden remains in effect on an element after it is relocated using aria-owns
